requirements.txt
Django==1.9.7
django-extensions==1.6.7
django-extras==0.3
django-solo==1.1.2
django-multiselectfield==0.1.3
